src/main/java/org/oop/Point.java
package org.oop;

public class Point {
    private final double x;
    private final double y;

    public double getX() {
        return x;
    }

    public double getY() {
        return y;
    }

    public Point(double x, double y) {
        this.x = x;
        this.y = y;
    }

    public double distanceFrom(Point anotherPoint) {
        double xDistance = this.getX() - anotherPoint.getX();
        double yDistance = this.getY() - anotherPoint.getY();
        return Math.sqrt(Math.pow(xDistance, 2) + Math.pow(yDistance, 2));
    }
}
